+++ NomeroffNet/Base/TextImageGenerator.py
@@ -76,11 +76,20 @@ class TextImageGenerator:
             ret.append(outstr)
         return ret
 
-    def build_data(self, aug_count=0):
+    def build_data(self, use_aug = False):
         self.imgs = np.zeros((self.n, self.img_h, self.img_w))
         self.texts = []
+        
+       
         for i, (img_filepath, text) in enumerate(self.samples):
             img = cv2.imread(img_filepath)
+            
+            if use_aug:
+                img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+                imgs = aug([img])
+                img = cv2.cvtColor(img, cv2.COLOR_RGB2BGR)
+                img = imgs[0]
+
             img = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
 
             # CLAHE
@@ -95,28 +104,7 @@ class TextImageGenerator:
             # because width is the time dimension when it gets fed into the RNN
             self.imgs[i, :, :] = img
             self.texts.append(text)
-        while aug_count:
-            for i, (img_filepath, text) in enumerate(self.samples):
-                img = cv2.imread(img_filepath)
-                img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-                imgs = aug([img])
-                img = cv2.cvtColor(img, cv2.COLOR_RGB2BGR)
-                img = imgs[0]
-                img = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-
-                # CLAHE
-                clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8,8))
-                img = clahe.apply(img)
-
-                img = cv2.resize(img, (self.img_w, self.img_h))
-                img = img.astype(np.float32)
-                img -= np.amin(img)
-                img /= np.amax(img)
-                # width and height are backwards from typical Keras convention
-                # because width is the time dimension when it gets fed into the RNN
-                self.imgs[i, :, :] = img
-                self.texts.append(text)
-            aug_count -= 1
+            
         self.n = len(self.imgs)
         self.indexes = list(range(self.n))
 
@@ -174,7 +162,7 @@ class TextImageGenerator:
                 else:
                     img = np.expand_dims(img, -1)
                 X_data[i] = img
-                Y_data[i] = self.text_to_labels(text)
+                Y_data[i] = np.array(self.text_to_labels(text))
                 source_str.append(text)
                 label_length[i] = len(text)