+++ NomeroffNet/Base/__init__.py
@@ -2,4 +2,3 @@ from .OCR import *
 from .TextImageGenerator import *
 from .ImgGenerator import *
 from .aug import *
-from .convert_keras_to_freeze_pb import *