+++ NomeroffNet/__init__.py
@@ -1,8 +1,7 @@
-from .mcm import *
-from .Detector import *
 from .filters import *
 from .RectDetector import *
 from .TextDetector import *
 from .TextPostprocessing import *
 from .OptionsDetector import *
-__version__ = "0.3.5"
+from .DetectronDetector import *
+__version__ = "0.4.0"