+++ examples/py/demo3.py
@@ -1,29 +1,25 @@
-# Import all necessary libraries.
+# Specify device
 import os
-import cv2
+os.environ["CUDA_VISIBLE_DEVICES"] = "" 
+
+# Import all necessary libraries.
 import numpy as np
 import sys
-import json
+import glob
 import matplotlib.image as mpimg
+import matplotlib.pyplot as plt
 
-# change this property
-NOMEROFF_NET_DIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), '../../')
-print(NOMEROFF_NET_DIR)
-
-# specify the path to Mask_RCNN if you placed it outside Nomeroff-net project
-MASK_RCNN_DIR = os.path.join(NOMEROFF_NET_DIR, 'Mask_RCNN')
-MASK_RCNN_LOG_DIR = os.path.join(NOMEROFF_NET_DIR, 'logs')
-
+# NomeroffNet path
+NOMEROFF_NET_DIR = os.path.abspath('../../')
 sys.path.append(NOMEROFF_NET_DIR)
 
 # Import license plate recognition tools.
-from NomeroffNet import  filters, RectDetector, TextDetector, OptionsDetector, Detector, textPostprocessing, textPostprocessingAsync
-
-print("LOADING MODELS...")
-
-# Initialize npdetector with default configuration file.
-nnet = Detector(MASK_RCNN_DIR, MASK_RCNN_LOG_DIR)
-nnet.loadModel("latest")
+from NomeroffNet import  filters
+from NomeroffNet import  RectDetector
+from NomeroffNet import  TextDetector
+from NomeroffNet import  OptionsDetector
+from NomeroffNet import  Detector
+from NomeroffNet import  textPostprocessing
 
 rectDetector = RectDetector()
 
@@ -36,8 +32,12 @@ textDetector = TextDetector({
         "for_regions": ["eu_ua_2015", "eu_ua_2004"],
         "model_path": "latest"
     },
+    "eu_ua_1995": {
+        "for_regions": ["eu_ua_1995"],
+        "model_path": "latest"
+    },
     "eu": {
-        "for_regions": ["eu", "eu_ua_1995"],
+        "for_regions": ["eu"],
         "model_path": "latest"
     },
     "ru": {
@@ -54,50 +54,41 @@ textDetector = TextDetector({
     }
 })
 
+# Initialize npdetector with default configuration file.
+nnet = Detector()
+nnet.loadModel(NOMEROFF_NET_DIR)
+
+
 # Walking through the ./examples/images/ directory and checking each of the images for license plates.
-print("START RECOGNIZING")
-rootDir = os.path.join(os.path.dirname(os.path.realpath(__file__)), '../images/')
+rootDir = '../images/*'
 
-max_img_w = 1600
-for dirName, subdirList, fileList in os.walk(rootDir):
-    for fname in fileList:
-        img_path = os.path.join(dirName, fname)
-        print(img_path)
-        img = mpimg.imread(img_path)
-         
-        # corect size for better speed
-        img_w = img.shape[1]
-        img_h = img.shape[0]
-        img_w_r = 1
-        img_h_r = 1
-        if img_w > max_img_w:
-            resized_img = cv2.resize(img, (max_img_w, int(max_img_w/img_w*img_h)))
-            img_w_r = img_w/max_img_w
-            img_h_r = img_h/(max_img_w/img_w*img_h)
-        else:
-            resized_img = img
-
-        NP = nnet.detect([resized_img]) 
+imgs = [mpimg.imread(img_path) for img_path in glob.glob(rootDir)]
         
-        # Generate image mask.
-        cv_img_masks = filters.cv_img_mask(NP)
+cv_imgs_masks = nnet.detect_mask(imgs)
+
+for img, cv_img_masks in zip(imgs, cv_imgs_masks):    
+    # Detect points.
+    arrPoints = rectDetector.detect(cv_img_masks)
             
-        # Detect points.
-        arrPoints = rectDetector.detect(cv_img_masks, outboundHeightOffset=0, fixGeometry=True, fixRectangleAngle=10)
-        print(arrPoints)
-        arrPoints[..., 1:2] = arrPoints[..., 1:2]*img_h_r
-        arrPoints[..., 0:1] = arrPoints[..., 0:1]*img_w_r
-        
-        # cut zones
-        zones = rectDetector.get_cv_zonesBGR(img, arrPoints)
-    
-        # find standart
-        regionIds, stateIds, countLines = optionsDetector.predict(zones)
-        regionNames = optionsDetector.getRegionLabels(regionIds)
-        print(regionNames)
-        print(countLines)
+    # Detect points.
+    arrPoints = rectDetector.detect(cv_img_masks)
 
-        # find text with postprocessing by standart  
-        textArr = textDetector.predict(zones, regionNames, countLines)
-        textArr = textPostprocessing(textArr, regionNames)
-        print(textArr)
+    # cut zones
+    zones = rectDetector.get_cv_zonesBGR(img, arrPoints)
+    toShowZones = rectDetector.get_cv_zonesRGB(img, arrPoints)
+    for zone, points in zip(toShowZones, arrPoints):
+        plt.axis("off")
+        plt.imshow(zone)
+        plt.show()
+
+    # find standart
+    regionIds, stateIds, countLines = optionsDetector.predict(zones)
+    regionNames = optionsDetector.getRegionLabels(regionIds)
+    print(regionNames)
+    print(countLines)
+
+    # find text with postprocessing by standart  
+    textArr = textDetector.predict(zones, regionNames, countLines)
+    textArr = textPostprocessing(textArr, regionNames)
+    print(textArr)
+